Fix calcualtion part 2

+++ topologies/multi.go
@@ -38,14 +38,15 @@ func (t *MultiTopology) IsRouteReflector(nodeID string, labels map[string]string
 	return ok
 }
 
-func (t *MultiTopology) GetClusterID(nodeID string, seed int) string {
+func (t *MultiTopology) GetClusterID(nodeID string, seed int64) string {
+	// TODO creation time
 	count := strings.Count(t.ClusterID, "%d")
 	parts := make([]interface{}, 0)
 
 	rand1 := rand.New(rand.NewSource(int64(getRouteReflectorID(nodeID))))
 	parts = append(parts, rand1.Int31n(254))
 
-	rand2 := rand.New(rand.NewSource(int64(seed)))
+	rand2 := rand.New(rand.NewSource(seed))
 	for len(parts) < count {
 		parts = append(parts, rand2.Int31n(254))
 	}
